Move userns cli test to a separate file, remove experimental flag
